refactor/ findOverlappedHighlight 로직 스토어로 이동
