+++ src/store/file/highlight.ts
@@ -1,6 +1,8 @@
 import { defineStore } from 'pinia';
 import { ref } from 'vue';
 import Highlight, { IHighlight } from '@/classes/Highlight';
+import { IPos } from '@/Interface/IPos';
+import getSelectedLines from '@/utils/getSelectedLines';
 
 export const useHighlightStore = defineStore('highlight', () => {
     const highlightList = ref<Highlight[]>([]);
@@ -16,6 +18,20 @@ export const useHighlightStore = defineStore('highlight', () => {
     function addHighlight(highlight: Highlight) {
         highlightList.value.push(highlight);
     }
+    function findOverlappedHighlight(
+        pos: IPos,
+        pageNum: number
+    ): Highlight | undefined {
+        const highlights = getHiglightsInPage(pageNum);
+
+        return highlights.find((h) => {
+            const range = h.getRange();
+            if (!range) return;
+
+            const lines = getSelectedLines(range);
+            return lines.some((line) => line.rect.isOverlap(pos.x, pos.y));
+        });
+    }
     function toJSON() {
         return highlightList.value.map((highlight) => highlight.toJSON());
     }
@@ -25,6 +41,7 @@ export const useHighlightStore = defineStore('highlight', () => {
     return {
         highlightList,
         getHiglightsInPage,
+        findOverlappedHighlight,
         deleteHighlight,
         addHighlight,
         toJSON,
